% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processMiningModule.R
\name{processMiningUI}
\alias{processMiningUI}
\alias{processMiningModule}
\title{Shiny module for process mining visualization and analysis of status development of registration}
\usage{
processMiningUI(id)

processMiningModule(
  input,
  output,
  session,
  .pm,
  audit_data = reactive(NULL),
  column = NULL,
  option_json = reactive(NULL)
)
}
\arguments{
\item{id}{Shiny input id}

\item{.reg}{formClass for which the process mining is being done (call to formClass does not have event_data set as NULL)}
}
\description{
Shiny module for process mining visualization and analysis of status development of registration
}
