% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmClass.R
\name{pmClass}
\alias{pmClass}
\title{R6 Class voor Process Mining applicaties}
\description{
R6 Class voor Process Mining applicaties

R6 Class voor Process Mining applicaties
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NA-new}{\code{pmClass$new()}}
\item \href{#method-NA-connect_to_database}{\code{pmClass$connect_to_database()}}
\item \href{#method-NA-close}{\code{pmClass$close()}}
\item \href{#method-NA-make_column}{\code{pmClass$make_column()}}
\item \href{#method-NA-read_table}{\code{pmClass$read_table()}}
\item \href{#method-NA-append_data}{\code{pmClass$append_data()}}
\item \href{#method-NA-table_columns}{\code{pmClass$table_columns()}}
\item \href{#method-NA-query}{\code{pmClass$query()}}
\item \href{#method-NA-has_value}{\code{pmClass$has_value()}}
\item \href{#method-NA-replace_value_where}{\code{pmClass$replace_value_where()}}
\item \href{#method-NA-make_choices}{\code{pmClass$make_choices()}}
\item \href{#method-NA-choices_from_json}{\code{pmClass$choices_from_json()}}
\item \href{#method-NA-from_json}{\code{pmClass$from_json()}}
\item \href{#method-NA-to_json}{\code{pmClass$to_json()}}
\item \href{#method-NA-add_event}{\code{pmClass$add_event()}}
\item \href{#method-NA-delete_events_from_case}{\code{pmClass$delete_events_from_case()}}
\item \href{#method-NA-make_event_data}{\code{pmClass$make_event_data()}}
\item \href{#method-NA-get_eventdata_case}{\code{pmClass$get_eventdata_case()}}
\item \href{#method-NA-get_auditdata_case}{\code{pmClass$get_auditdata_case()}}
\item \href{#method-unknown-clone}{\code{pmClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-new"></a>}}
\if{latex}{\out{\hypertarget{method-NA-new}{}}}
\subsection{Method \code{new()}}{
Make new process mining application object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmClass$new(
  config_file = "conf/config.yml",
  what,
  schema = NULL,
  pool = TRUE,
  sqlite = NULL,
  form_data = NULL,
  form_audit = NULL,
  event_data = NULL,
  event_columns = list(case = "case_id", activity = "activity_id", activity_instance =
    "activity_id_instance", eventtime = "eventtime", lifecycle = "lifecycle_id", resource
    = "resource")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config_file}}{Path to DB config}

\item{\code{what}}{Entry in config for DB connection}

\item{\code{schema}}{DB schema}

\item{\code{pool}}{If TRUE, connects with dbPool}

\item{\code{sqlite}}{If path to an SQLite file, uses SQLite.}

\item{\code{form_data}}{If the user uses the shintoforms package this table is used to get the current state}

\item{\code{form_audit}}{If the user uses the shintoforms package this table is used to get the history}

\item{\code{event_data}}{If the user want to use their own event data}

\item{\code{event_columns}}{If the user uses own event_data, the columns can be specified here.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-connect_to_database"></a>}}
\if{latex}{\out{\hypertarget{method-NA-connect_to_database}{}}}
\subsection{Method \code{connect_to_database()}}{
Connect to a database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmClass$connect_to_database(
  config_file = NULL,
  schema = NULL,
  what = NULL,
  pool = TRUE,
  sqlite = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-close"></a>}}
\if{latex}{\out{\hypertarget{method-NA-close}{}}}
\subsection{Method \code{close()}}{
Close database connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmClass$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-make_column"></a>}}
\if{latex}{\out{\hypertarget{method-NA-make_column}{}}}
\subsection{Method \code{make_column()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmClass$make_column(table, column, type = "varchar")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-read_table"></a>}}
\if{latex}{\out{\hypertarget{method-NA-read_table}{}}}
\subsection{Method \code{read_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmClass$read_table(table, lazy = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-append_data"></a>}}
\if{latex}{\out{\hypertarget{method-NA-append_data}{}}}
\subsection{Method \code{append_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmClass$append_data(table, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-table_columns"></a>}}
\if{latex}{\out{\hypertarget{method-NA-table_columns}{}}}
\subsection{Method \code{table_columns()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmClass$table_columns(table)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-query"></a>}}
\if{latex}{\out{\hypertarget{method-NA-query}{}}}
\subsection{Method \code{query()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmClass$query(txt, glue = TRUE, quiet = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-has_value"></a>}}
\if{latex}{\out{\hypertarget{method-NA-has_value}{}}}
\subsection{Method \code{has_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmClass$has_value(table, column, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-replace_value_where"></a>}}
\if{latex}{\out{\hypertarget{method-NA-replace_value_where}{}}}
\subsection{Method \code{replace_value_where()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmClass$replace_value_where(
  table,
  col_replace,
  val_replace,
  col_compare,
  val_compare,
  query_only = FALSE,
  quiet = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-make_choices"></a>}}
\if{latex}{\out{\hypertarget{method-NA-make_choices}{}}}
\subsection{Method \code{make_choices()}}{
Make choices (for selectInput) based on values and names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmClass$make_choices(
  values_from,
  names_from = values_from,
  data = NULL,
  sort = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-choices_from_json"></a>}}
\if{latex}{\out{\hypertarget{method-NA-choices_from_json}{}}}
\subsection{Method \code{choices_from_json()}}{
Unpack a JSON field to make a named vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmClass$choices_from_json(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-from_json"></a>}}
\if{latex}{\out{\hypertarget{method-NA-from_json}{}}}
\subsection{Method \code{from_json()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmClass$from_json(x, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-NA-to_json}{}}}
\subsection{Method \code{to_json()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmClass$to_json(x, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-add_event"></a>}}
\if{latex}{\out{\hypertarget{method-NA-add_event}{}}}
\subsection{Method \code{add_event()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmClass$add_event(
  case_id,
  activity,
  resource,
  act_ins = NULL,
  e_time = NULL,
  lc = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-delete_events_from_case"></a>}}
\if{latex}{\out{\hypertarget{method-NA-delete_events_from_case}{}}}
\subsection{Method \code{delete_events_from_case()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmClass$delete_events_from_case(case_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-make_event_data"></a>}}
\if{latex}{\out{\hypertarget{method-NA-make_event_data}{}}}
\subsection{Method \code{make_event_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmClass$make_event_data(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_eventdata_case"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_eventdata_case}{}}}
\subsection{Method \code{get_eventdata_case()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmClass$get_eventdata_case(id, column, option_json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_auditdata_case"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_auditdata_case}{}}}
\subsection{Method \code{get_auditdata_case()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmClass$get_auditdata_case(id, column, option_json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unknown-clone"></a>}}
\if{latex}{\out{\hypertarget{method-unknown-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
